Refactor reward services

+++ apps/api/src/app/controllers/pools/quests/delete.controller.ts
@@ -1,14 +1,21 @@
-import { body, param } from 'express-validator';
+import { param } from 'express-validator';
 import { Request, Response } from 'express';
-import { serviceMap } from '@thxnetwork/api/services/interfaces/IQuestService';
+import { ForbiddenError } from '@thxnetwork/api/util/errors';
+import { QuestVariant } from '@thxnetwork/common/enums';
 import LockService from '@thxnetwork/api/services/LockService';
+import QuestService from '@thxnetwork/api/services/QuestService';
 
-const validation = [param('id').isMongoId(), body('variant').isInt()];
+const validation = [param('id').isMongoId(), param('variant').isInt(), param('questId').isMongoId()];
 
 const controller = async (req: Request, res: Response) => {
+    const variant = req.params.variant as unknown as QuestVariant;
+    const poolId = req.params.id;
     const questId = req.params.questId;
-    const Quest = serviceMap[req.body.variant].models.quest;
-    await Quest.findByIdAndRemove(questId);
+
+    const quest = await QuestService.findById(variant, questId);
+    if (quest.poolId !== poolId) throw new ForbiddenError('Not your quest.');
+
+    await quest.deleteOne();
 
     // Remove all locks for this quest
     await LockService.removeAllLocks(questId);
